% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr-package.R
\docType{package}
\name{bpr-package}
\alias{bpr}
\alias{bpr-package}
\title{bpr: Fitting Bayesian Poisson Regression}
\description{
Posterior sampling and inference for Bayesian Poisson regression models. The model specification makes use of Gaussian (or conditionally Gaussian) prior distributions on the regression coefficients. Details on the algorithm are found in D'Angelo and Canale (2021) \href{https://arxiv.org/abs/2109.09520}{arXiv:2109.09520}.
}
\keyword{internal}
