% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{mcmc_diagnostics.poisreg}
\alias{mcmc_diagnostics.poisreg}
\title{Print MCMC Convergence Diagnostics}
\usage{
\method{mcmc_diagnostics}{poisreg}(object, perc_burnin = NULL)
}
\arguments{
\item{object}{object of class "\code{poisreg}" (usually, the result of a call to \code{\link{sample_bpr}}).}

\item{perc_burnin}{(optional) percentage of each chain to be discarded as burn-in.}
}
\value{
\code{mcmc_diagnostics.poisreg} returns a list with elements:

\code{chain_length} total length of the MCMC chains.

\code{len_burnin}  the length of the burn-in used to compute the estimates.

\code{thin} the thinning frequency used (from \code{object}).

\code{effSize} effective sample size of each parameter chain after removing burn-in and thinning.

\code{geweke} Geweke diagnostics of convergence of the chains (value of the test and p-value).

\code{gelman_rubin} if \code{nchains > 1}, Gelman-Rubin diagnostics of convergence.
}
\description{
Print MCMC Convergence Diagnostics
}
