% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics_export.R
\name{mcmc_diagnostics}
\alias{mcmc_diagnostics}
\title{MCMC Convergence Diagnostics}
\usage{
mcmc_diagnostics(object)
}
\arguments{
\item{object}{an object for which mcmc diagnostic is desired.}
}
\description{
MCMC Convergence Diagnostics
}
\seealso{
\code{\link{mcmc_diagnostics.poisreg}}
}
