% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.poisreg}
\alias{merge.poisreg}
\title{Merge Multiple Chains}
\usage{
\method{merge}{poisreg}(object, perc_burnin = NULL)
}
\arguments{
\item{object}{object of class "\code{poisreg}" (usually, the result of a call to \code{\link{sample_bpr}}), with \code{nchains > 1}.}

\item{perc_burnin}{(optional) percentage of each chain to be discarded as burn-in.}
}
\value{
The function returns an object of class \code{poisreg} with a single element \code{$sim}. 
The returned chains (elements of \code{sim}) are obtained by appending the simulated values of each independent chain, 
under the assumption that they all have reached the same stationary distribution.
}
\description{
This function is a method for class \code{poisreg}. Merge multiple MCMC chains into a unique chain.
}
\examples{
## For examples, see example(sample_bpr).
}
