% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_export.R
\name{merge_sim}
\alias{merge_sim}
\title{Merge Multiple Chains}
\usage{
merge_sim(object)
}
\arguments{
\item{object}{object}
}
\description{
Merge Multiple Chains
}
\seealso{
\code{\link{merge_sim.poisreg}}
}
