% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.poisreg}
\alias{plot.poisreg}
\title{Plot Trace and Distribution of Regression Parameters}
\usage{
\method{plot}{poisreg}(x, ...)
}
\arguments{
\item{x}{object of class "\code{poisreg}" (usually, the result of a call to \code{\link{sample_bpr}}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot Trace and Distribution of Regression Parameters
}
\seealso{
\code{\link{sample_bpr}}
}
