% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppc.R
\name{plot.posterior_check}
\alias{plot.posterior_check}
\title{Plot posterior check}
\usage{
\method{plot}{posterior_check}(object, perc_burnin = NULL, ...)
}
\arguments{
\item{object}{}

\item{perc_burnin}{}

\item{...}{}
}
\description{
Plot posterior check
}
