% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{posterior_predictive.poisreg}
\alias{posterior_predictive.poisreg}
\title{Title}
\usage{
\method{posterior_predictive}{poisreg}(object, perc_burnin = NULL, newdata = NULL)
}
\arguments{
\item{object}{}

\item{perc_burnin}{}

\item{newdata}{}
}
\value{

}
\description{
Title
}
