% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_export.R
\name{posterior_predictive}
\alias{posterior_predictive}
\title{Compute Posterior Predictive Distribution}
\usage{
posterior_predictive(object, new_X)
}
\arguments{
\item{object}{object of class "\code{poisreg}" (usually, the result of a call to \code{\link{sample_bpr}}).}

\item{new_X}{(optional) a data frame in which to look for variables with which to predict.}
}
\description{
Compute Posterior Predictive Distribution
}
\seealso{
\code{\link{posterior_predictive.poisreg}}
}
