% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_bpr}
\alias{sample_bpr}
\title{Fitting Bayesian Poisson Regression}
\usage{
sample_bpr(
  formula = NULL,
  data = NULL,
  y = NULL,
  X = NULL,
  iter,
  burnin = NULL,
  prior = list(type = "gaussian", b = NULL, B = NULL, tau = NULL),
  pars = list(method = "MH", max_dist = 50, max_r = NULL, max_dist_burnin = 1e+06),
  state = NULL,
  thin = 1,
  verbose = TRUE,
  seed = NULL,
  nchains = 1
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{data frame or matrix containing the variables in the model.}

\item{y, X}{response variable and matrix of covariates. Alternatively to providing formula and data, one can pass y and X.}

\item{iter}{number of algorithm iterations.}

\item{burnin}{(optional) a positive integer specifying the length of the burn-in. 
If a value > 1 is provided, the first \code{burnin} iterations use a different tuning parameter in order to better explore the parameters space.}

\item{prior}{a named list of parameters to select prior type and parameters. See the details for the \code{prior} arguments.}

\item{pars}{a named list of parameters to select algorithm type and tuning parameters. See the details for the \code{pars} arguments.}

\item{state}{optional vector providing the starting points for the chains.}

\item{thin}{a positive integer specifying the period for saving samples. The default is 1.}

\item{verbose}{logical (default = TRUE) indicating whether to print messages on the progress of the algorithm and possible convergence issues.}

\item{seed}{(optional) positive integer: the seed of random number generator.}

\item{nchains}{(optional) positive integer specifying the number of Markov chains. The default is 1.}
}
\value{
An object of S3 class \code{poisreg} containing the results of the sampling. \cr\cr
The function \code{\link{summary}} can be used to print a summary of the posterior inference for the regression parameters and of the algorithm diagnostics. \cr
The function \code{\link{mcmc_diagnostics}} prints convergence diagnostics for the sampled chains.  \cr
The function \code{plot} prints the trace of the sampled values and a density estimate of the regression coefficients. See \code{\link[coda]{plot.mcmc}}.\cr
The function \code{\link{posterior_predictive}} allows to compute the posterior predictive distributions to check the model. See also the related function \code{\link{plot.ppc}}.

\code{poisreg} is a list containing at least the following elements:

\code{sim} : list of the results of the sampling. It contains the following elements: \itemize{
\item{ \code{beta} : \code{\link[coda]{mcmc}} object of posterior draws of the regression coefficients.}
\item{ \code{r} : the sequence of adaptive tuning parameters used in each iteration. }
\item{ \code{time} : the total amount of time to perform the simulation. }
}

\code{formula}  : the \code{formula} object used.

\code{data}  : list of the data (vector of counts and covariate matrix).

\code{state}  : the starting points of the chain.

\code{burnin}  : length of the used \code{burnin}.

\code{prior}  : whether a Gaussian or horseshoe prior was used.

\code{prior_pars}  : prior parameters.

\code{thin}  : thinning frequency passed to the \code{thin} parameter.

\code{nchains}  : number of chains. If \code{nchains} was chosen >1, the list will also include additional numbered \code{sim} elements, one for each sampled chain.
}
\description{
\code{sample_bpr} is used to generate draws from the posterior distribution of the coefficients of Poisson regression models. 
The method allows for Gaussian and horseshoe (Carvalho et al, 2010) prior distributions.
}
\examples{
# Minimal call
out = sample_bpr(y = y, X = X, iter = 2000)
summary(out)   # summary of posterior inference
mcmc_diagnostics(out)   # summary of MCMC convergence diagnostics
plot(out)   # plot the traceplot and posterior density of the regression coefficients

}
\references{
Carvalho, C., Polson, N., & Scott, J. (2010). The horseshoe estimator for sparse signals. Biometrika, 97(2), 465-480.
}
