% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Title}
\usage{
sampling(
  formula,
  data,
  y,
  X,
  iter,
  burnin,
  prior,
  pars,
  r_start,
  state,
  thin,
  verbose,
  seed,
  trunc_lambda
)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{y}{}

\item{X}{}

\item{iter}{}

\item{burnin}{}

\item{prior}{}

\item{pars}{}

\item{r_start}{}

\item{state}{}

\item{thin}{}

\item{verbose}{}

\item{seed}{}

\item{trunc_lambda}{}
}
\value{
list
}
\description{
Title
}
\examples{

}
