% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.poisreg}
\alias{summary.poisreg}
\title{Print Summary of the output of sample_bpr()}
\usage{
\method{summary}{poisreg}(object, perc_burnin = NULL)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Print Summary of the output of sample_bpr()
}
