% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_bpr}
\alias{sample_bpr}
\title{Fitting Bayesian Poisson Regression}
\usage{
sample_bpr(
  formula = NULL,
  data = NULL,
  iter,
  burnin = NULL,
  prior = list(type = "gaussian", b = NULL, B = NULL, tau = NULL),
  pars = list(method = "MH", max_dist = 50, max_r = NULL, max_dist_burnin = 1e+06),
  state = NULL,
  thin = 1,
  verbose = TRUE,
  seed = NULL,
  nchains = 1,
  perc_burnin = 0.25
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{data frame or matrix containing the variables in the model.}

\item{iter}{number of algorithm iterations.}

\item{burnin}{(optional) a positive integer specifying the length of the burn-in. 
If a value > 1 is provided, the first \code{burnin} iterations use a different tuning parameter in order to better explore the parameter space.}

\item{prior}{a named list of parameters to select prior type and parameters. See the details for the \code{prior} arguments.}

\item{pars}{a named list of parameters to select algorithm type and tuning parameters. See the details for the \code{pars} arguments.}

\item{state}{optional vector providing the starting points of the chains.}

\item{thin}{a positive integer specifying the period for saving samples. The default is 1.}

\item{verbose}{logical (default = TRUE) indicating whether to print messages on the progress of the algorithm and possible convergence issues.}

\item{seed}{(optional) positive integer: the seed of random number generator.}

\item{nchains}{(optional) positive integer specifying the number of Markov chains. The default is 1.}

\item{perc_burnin}{(default = 0.25) percentage of the chain to be discarded to perform inference. If both burnin and perc_burnin are specified, the most conservative burn-in is considered.}
}
\value{
An object of S3 class \code{poisreg} containing the results of the sampling. \cr
\code{poisreg} is a list containing at least the following elements:

\code{sim} : list of the results of the sampling. It contains the following elements: \itemize{
\item{ \code{beta} : \code{\link[coda]{mcmc}} object of posterior draws of the regression coefficients.}
\item{ \code{r} : the sequence of adaptive tuning parameters used in each iteration. }
\item{ \code{time} : the total amount of time to perform the simulation. }
}

\code{formula}  : the \code{formula} object used.

\code{data}  : list with elements the matrix of covariates \code{X} and response variable \code{y}.

\code{state}  : the starting points of the chain.

\code{burnin}  : length of the used burn-in.

\code{prior}  : whether a Gaussian or horseshoe prior was used.

\code{prior_pars}  : prior parameters.

\code{thin}  : thinning frequency passed to the \code{thin} parameter.

\code{nchains}  : number of chains. If \code{nchains} was chosen >1, the output list will also include additional 
numbered \code{sim} elements, one for each sampled chain.

\code{perc_burnin} : percentage of the chain used as burn-in.
}
\description{
\code{sample_bpr} is used to generate draws from the posterior distribution of the coefficients of Poisson regression models. 
The method allows for Gaussian and horseshoe (Carvalho et al, 2010) prior distributions, 
and relies on a Metropolis-Hastings or importance sampler algorithm. See D'Angelo and Canale (2021) for details on the algorithms.
}
\details{
An object of class \code{poisreg} admits class-specific methods to analyze the output.\cr
The function \code{\link{summary.poisreg}} can be used to obtain or print a summary of the results and of the algorithm diagnostics. \cr
The function \code{\link{mcmc_diagnostics.poisreg}} can be used to obtain or print convergence diagnostics for the sampled chains.  \cr
The function \code{\link{plot.poisreg}} prints the trace of the sampled values and a density estimate of the regression coefficients. 
See \code{\link[coda]{plot.mcmc}}.\cr
The function \code{\link{posterior_predictive.poisreg}} can be used to compute the posterior predictive distributions to check the model. 
See also the related function \code{plot.ppc}.
}
\examples{
require(MASS) # load the data set
head(epil)

fit = sample_bpr( y ~  lbase*trt + lage + V4, data = epil, 
                   iter = 1000)
                   
summary(fit)    # summary of posterior inference
mcmc_diagnostics(fit)    # summary of MCMC convergence diagnostics

plot(fit)    


## Examples with different options
# Select prior parameters and set tuning parameter 
fit2 = sample_bpr( y ~  lbase*trt + lage + V4, data = epil, 
                    iter = 1000, 
                    prior = list( type = "gaussian", b = rep(0, 6), 
                                  B = diag(6) * 3 ),
                    pars = list( max_dist = 10 ))
                    
# Horseshoe prior, with global shrinkage parameter equal to 1
fit3 = sample_bpr( y ~  lbase*trt + lage + V4, data = epil, 
                    iter = 1000, 
                    prior = list( type = "horseshoe", tau = 1 ),
                    pars = list( max_dist = 10 ))     
                    
# Simulate multiple chains and merge outputs after checking convergence
fit4 = sample_bpr( y ~  lbase*trt + lage + V4, data = epil, 
                    iter = 1000, 
                    nchains = 4, thin = 2)
# fit now contains additional elements ($sim2, $sim3, $sim4)

mcmc_diagnostics(fit4)     
# the Gelman-Rubin diagnostics confirms convergence of the 4 
# independent chains to the same stationary distribution

fit4b = merge_sim(fit4) 
str(fit4b$sim)    
# fit 4b contains only one MCMC chain of length 1500 
# (after thinning and burn-in)

}
\references{
Carvalho, C., Polson, N., & Scott, J. (2010). The horseshoe estimator for sparse signals. Biometrika, 97(2), 465-480.\cr\cr
D'Angelo, L., and Canale, A. (2021) Efficient posterior sampling for Bayesian Poisson regression. arXiv...
}
\seealso{
\code{\link{summary.poisreg}} , \code{\link{mcmc_diagnostics.poisreg}} , \code{\link{plot.poisreg}} ,
\code{\link{merge_sim.poisreg}} , \code{\link{posterior_predictive.poisreg}}
}
